+++ android/src/main/kotlin/hamza/dali/flutter_osm_plugin/overlays/CustomLocation.kt
@@ -87,8 +87,24 @@ class CustomLocationManager(mapView: MapView) : MyLocationNewOverlay(mapView) {
         super.onLocationChanged(location, source)
     }
 
+    fun toggleFollow(enableStop: Boolean) {
+        enableAutoStop = enableStop
+        when (enableStop) {
+            true -> disableFollowLocation()
+            else -> {}
+        }
+    }
+
     fun followLocation(onChangedLocation: (gp: GeoPoint) -> Unit) {
         this.enableFollowLocation()
+        runOnFirstFix {
+            val location = this.lastFix
+            val geoPMap = GeoPoint(location)
+            onChangedLocation(geoPMap)
+        }
+    }
+
+    fun onChangedLocation(onChangedLocation: (gp: GeoPoint) -> Unit) {
         runOnFirstFix {
             val location = this.lastFix
             val geoPMap = GeoPoint(location)
@@ -104,6 +120,7 @@ class CustomLocationManager(mapView: MapView) : MyLocationNewOverlay(mapView) {
                     drawOnlyPerson(canvas, pProjection, lastFix)
                 }
             }
+
             else -> super.draw(canvas, pProjection)
         }
 
@@ -122,6 +139,7 @@ class CustomLocationManager(mapView: MapView) : MyLocationNewOverlay(mapView) {
 //                    mScale * (personIcon.width / 3f) + 0.5f,
 //                )
             }
+
             personIcon != null -> {
                 setPersonIcon(personIcon)
             }