+++ example/lib/src/simple_example_hook.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_osm_plugin/flutter_osm_plugin.dart';
-import 'package:osm_flutter_hooks/osm_flutter_hooks.dart';
+//import 'package:osm_flutter_hooks/osm_flutter_hooks.dart';
 
 class SimpleExample extends StatefulWidget {
   SimpleExample({Key? key}) : super(key: key);
@@ -35,15 +35,15 @@ class _SimpleExampleState extends State<SimpleExample> {
 class SimpleOSM extends HookWidget {
   @override
   Widget build(BuildContext context) {
-    final controller = useMapController(initMapWithUserPosition: true);
-    useMapIsReady(
-      controller: controller,
-      mapIsReady: () async {
-        await controller.setZoom(zoomLevel: 15);
-      },
-    );
+    // final controller = useMapController(initMapWithUserPosition: true);
+    // useMapIsReady(
+    //   controller: controller,
+    //   mapIsReady: () async {
+    //     await controller.setZoom(zoomLevel: 15);
+    //   },
+    // );
     return OSMFlutter(
-      controller: controller,
+      controller: MapController(),
       markerOption: MarkerOption(
         defaultMarker: MarkerIcon(
           icon: Icon(
@@ -53,7 +53,6 @@ class SimpleOSM extends HookWidget {
           ),
         ),
       ),
-      trackMyPosition: false,
     );
   }
 }