+++ flutter_osm_interface/lib/src/types/user_option.dart
@@ -1,0 +1,10 @@
+class UserTrackingOption {
+  final bool initWithUserPosition;
+  final bool enableTracking;
+  final bool unFollowUser;
+
+  const UserTrackingOption({
+    this.enableTracking = false,
+    this.unFollowUser = false,
+  }) : this.initWithUserPosition = true;
+}